requirements.txt
-r requirements/base.txt # includes the base.txt requirements file

psycopg2-binary==2.7.6.1
dj-database-url==0.5.0
django-heroku==0.3.1
django-oauth-toolkit==1.2.0
djangorestframework==3.9.0
gunicorn==19.9.0
requests==2.21.0
whitenoise==4.1.2
django-storages==1.7.1
boto3==1.9.71
botocore==1.12.71
s3transfer==0.1.13
docutils==0.14
jmespath==0.9.3
urllib3==1.24.1
six==1.12.0
python-dateutil==2.7.5